added support for sounds
now it has the original music in the background and i found some other sounds for the weapons and explosion. There is an issue with the ground still - when you fire it seems to offset the ground when there is a crater/hole. So I think the way the ground is drawn will need to be redone. The rocks are funky too, they jiggle a bit.